+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to fwtv module will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 
+## [2.2.0] - 2023-08-07
+
+## Added
+
+- extra window showing all attendances that lead to this verification failure
+- tolerance can be configured though input now
+
 ## [2.1.0] - 2023-04-08
 
 ### Changed